Add readme file and increase supported media types

+++ app/worker.py
@@ -1,6 +1,12 @@
 from urllib.parse import urlparse, parse_qs
 from telethon import TelegramClient
-from telethon.tl.types import DocumentAttributeVideo
+from telethon.tl.types import (
+    Message,
+    DocumentAttributeVideo,
+    DocumentAttributeAudio,
+    MessageMediaDocument,
+    MessageMediaPhoto,
+)
 import os
 from typing import Optional, Union
 
@@ -24,7 +30,7 @@ class Telegram_URL:
     topic_id: Optional[int] = None
 
 
-def check_url_type(telegram_url: str):
+def check_url_type(telegram_url: str) -> Telegram_URL:
 
     try:
         url_components = urlparse(telegram_url)
@@ -59,22 +65,40 @@ def check_url_type(telegram_url: str):
     return url_data
 
 
-async def get_group_message(user_client: TelegramClient, url_data: Telegram_URL):
+async def get_group_message(
+    user_client: TelegramClient, url_data: Telegram_URL
+) -> list:
     messages = await user_client.get_messages(
         url_data.channel_id,
         ids=[
             msg_id
-            for msg_id in range(url_data.message_id - 9, url_data.message_id + 10)
+            for msg_id in range(url_data.message_id - 9, url_data.message_id + 11)
         ],
     )
     media_group_id = (
         messages[9].grouped_id
-        if len(messages) == 19
+        if len(messages) == 20
         else messages[url_data.message_id - 1].grouped_id
     )
     if media_group_id is None:
         return list()
-    return list(msg for msg in messages if msg.grouped_id == media_group_id)
+    return list(msg for msg in messages if msg and msg.grouped_id == media_group_id)
+
+
+def get_media_type(message: Message) -> Optional[str]:
+    if not message.media:
+        return None
+    if isinstance(message.media, MessageMediaPhoto):
+        return "photo"
+    if isinstance(message.media, MessageMediaDocument):
+        doc = message.media.document
+        for attr in doc.attributes:
+            if hasattr(attr, "voice") and isinstance(attr.voice, bool):
+                return "voice" if attr.voice else "audio"
+            if hasattr(attr, "round_message") and isinstance(attr.round_message, bool):
+                return "video_note" if attr.round_message else "video"
+        return "document"
+    return None
 
 
 async def url_handler(
@@ -82,20 +106,12 @@ async def url_handler(
 ):
     url_data = check_url_type(telegram_url)
 
-    user_me = await user_client.get_me()
-    if not url_data.is_protect:
-        bot_read_msg_entiy = await bot_client.get_messages(
-            url_data.channel_id, ids=url_data.message_id
-        )
-        if bot_read_msg_entiy != None:
-            await bot_client.forward_messages(user_me.id, bot_read_msg_entiy)
-            return
-
     msg_entity = await user_client.get_messages(
         url_data.channel_id, ids=url_data.message_id
     )
     caption = msg_entity.message
     msg_media = getattr(msg_entity, "media")
+    user_me = await user_client.get_me()
     if not msg_media:
         await bot_client.send_message(user_me.id, caption)
         return
@@ -107,42 +123,67 @@ async def url_handler(
     if len(messages) <= 0:
         raise Exception("not get any message.")
 
+    if not url_data.is_protect:
+        bot_read_msg_entitys = await bot_client.get_messages(
+            url_data.channel_id, ids=list(msg.id for msg in messages if msg)
+        )
+        bot_read_msg_entitys = list(
+            msg_entitys for msg_entitys in bot_read_msg_entitys if msg_entitys
+        )
+        if len(bot_read_msg_entitys) > 0:
+            try:
+                await bot_client.forward_messages(user_me.id, bot_read_msg_entitys)
+                return
+            except:
+                pass
+
     files = []
     attributes = []
-    thumbs = []
+    voice = None
     for message in messages:
-        message_media = getattr(message, "media")
-        if not message_media:
+
+        media_type = get_media_type(message)
+        if not media_type:
             continue
 
-        if getattr(message_media, "video", None):
+        if media_type == "video":
             msg_file_metadata = getattr(message, "file")
-            height, width, duration = (
-                msg_file_metadata.height,
-                msg_file_metadata.width,
-                msg_file_metadata.duration,
-            )
             attributes.append(
                 DocumentAttributeVideo(
-                    h=height,
-                    w=width,
-                    duration=duration,
+                    h=msg_file_metadata.height,
+                    w=msg_file_metadata.width,
+                    duration=msg_file_metadata.duration,
                     round_message=True,
                     supports_streaming=True,
                 )
             )
-            thumbs.append(msg_file_metadata.media.thumbs)
-        else:
+        if media_type == "photo" or media_type == "document":
             attributes.append(None)
-
+        if media_type == "voice" or media_type == "audio":
+            msg_file_metadata = getattr(message, "file")
+            duration, title = msg_file_metadata.duration, msg_file_metadata.title
+            attributes.append(
+                DocumentAttributeAudio(
+                    duration=duration,
+                    voice=(media_type == "voice"),
+                    title=title,
+                )
+            )
         files.append(await user_client.download_media(message))
+        if media_type == "voice":
+            voice = files[len(files) - 1]
+            break
 
-    await bot_client.send_file(
-        user_me.id, files, caption=caption, attributes=attributes
-    )
-    for file in files:
-        os.remove(file)
-    print("done.")
+    if len(files) > 0:
+        wait_send_files = files
+        if voice != None:
+            wait_send_files = voice
+        await bot_client.send_file(
+            user_me.id, wait_send_files, caption=caption, attributes=attributes
+        )
+        for file in files:
+            os.remove(file)
+    print(f"done for url={telegram_url}")
 
 
 async def event_handler(event, user_client, bot_client):
